# Use memcmp to optimize == for bit integer arrays