--- a/base/array.jl
+++ b/base/array.jl
@@ -879,6 +879,18 @@ function lexcmp(a::Array{UInt8,1}, b::Array{UInt8,1})
     return c < 0 ? -1 : c > 0 ? +1 : cmp(length(a),length(b))
 end
 
+# use memcmp for == on bit integer types
+=={T<:BitInteger,N}(a::Array{T,N}, b::Array{T,N}) =
+    size(a) == size(b) &&
+    ccall(:memcmp, Int32, (Ptr{T}, Ptr{T}, UInt), a, b, sizeof(T) * length(a)) == 0
+
+# this is ~20% faster than the generic implementation above for very small arrays
+function =={T<:BitInteger}(a::Array{T,1}, b::Array{T,1})
+    len = length(a)
+    len == length(b) && ccall(
+        :memcmp, Int32, (Ptr{T}, Ptr{T}, UInt), a, b, sizeof(T) * len) == 0
+end
+
 function reverse(A::AbstractVector, s=1, n=length(A))
     B = similar(A)
     for i = 1:s-1