--- a/base/strings/basic.jl
+++ b/base/strings/basic.jl
@@ -157,11 +157,7 @@ isless(a::AbstractString, b::AbstractString) = cmp(a,b) < 0
 cmp(a::String, b::String) = lexcmp(a.data, b.data)
 cmp(a::Symbol, b::Symbol) = Int(sign(ccall(:strcmp, Int32, (Cstring, Cstring), a, b)))
 
-function ==(a::String, b::String)
-    len = length(a.data)
-    return len == length(b.data) &&
-        0 == ccall(:memcmp, Int32, (Ptr{UInt8}, Ptr{UInt8}, UInt), a.data, b.data, len)
-end
+==(a::String, b::String) = a.data == b.data
 isless(a::Symbol, b::Symbol) = cmp(a,b) < 0
 
 ## Generic validation functions ##