After https://github.com/JuliaLang/julia/pull/16855, it's faster to compare two arrays for equality with `String(a) == String(b)` than with `a == b`. No joke!

```julia
julia> using BenchmarkTools

julia> const A = repeat([0x01:0x7F;], outer=10000);

julia> const B = repeat([0x01:0x7F;], outer=10000);

julia> @benchmark A == B
BenchmarkTools.Trial: 
  samples:          2271
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     1.99 ms (0.00% GC)
  median time:      2.27 ms (0.00% GC)
  mean time:        2.20 ms (0.00% GC)
  maximum time:     2.46 ms (0.00% GC)

julia> @benchmark String(A) == String(B)
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  32.00 bytes
  allocs estimate:  2
  minimum time:     76.32 μs (0.00% GC)
  median time:      80.60 μs (0.00% GC)
  mean time:        82.95 μs (0.00% GC)
  maximum time:     314.34 μs (0.00% GC)
```

In fact, the fastest (non-String) way to compare byte arrays for equality seems to be `lexcmp`, which uses `memcmp` behind the hood:

```julia
julia> @benchmark lexcmp(A, B) == 0
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     76.49 μs (0.00% GC)
  median time:      86.93 μs (0.00% GC)
  mean time:        89.47 μs (0.00% GC)
  maximum time:     450.16 μs (0.00% GC)
```

I think it makes sense to move this optimization up one level, from strings to arrays. I'm not entirely sure what the scope should be—currently it's for one-dimensional arrays of integral types that are at most 64 bits and not `Bool`, which is a bit arbitrary. It doesn't work for floating point types because of `NaN` behaviour, and I'm a little unclear about the semantics of `Bool` in Julia, but I think it might work for that also, if safe.