@timholy 2016-06-11
Wow.

For, me, the following julia implementation:
```jl
function iseq3(A::AbstractArray, B::AbstractArray)
    if size(A) != size(B)
        return false
    end
    if isa(A,Range) != isa(B,Range)
        return false
    end
    failures = 0
    @simd for I in eachindex(A)
        @inbounds failures += A[I] != B[I]
    end
    return failures == 0
end
```
is (for the arrays you were testing on) much better than our current implementation (by 3-4x), but the `ccall` is another 4x better than that.