@TotalVerb 2016-06-11
Issues addressed. This final version, on my system, has the following benchmarking attributes:

```
NEW ==
rand(UInt8, 0): 8
rand(UInt8, 1): 7
rand(UInt8, 2): 7
rand(UInt8, 3): 7
rand(UInt8, 4): 8
rand(UInt8, 5): 7
rand(UInt8, 10): 8
rand(UInt8, 20): 7
rand(UInt8, 50): 9
rand(UInt8, 100): 11
rand(UInt8, 1000): 36
rand(UInt8, 10000): 292
rand(UInt8, 100000): 3552
rand(UInt8, 1000000): 56017
rand(UInt8, 10000000): 1003760

OLD ==
rand(UInt8, 0): 8
rand(UInt8, 1): 9
rand(UInt8, 2): 10
rand(UInt8, 3): 11
rand(UInt8, 4): 13
rand(UInt8, 5): 15
rand(UInt8, 10): 22
rand(UInt8, 20): 38
rand(UInt8, 50): 93
rand(UInt8, 100): 171
rand(UInt8, 1000): 1583
rand(UInt8, 10000): 15696
rand(UInt8, 100000): 156832
rand(UInt8, 1000000): 1566205
rand(UInt8, 10000000): 15829086

NEW ==
rand(Int64, 0): 7
rand(Int64, 1): 6
rand(Int64, 2): 6
rand(Int64, 3): 7
rand(Int64, 4): 7
rand(Int64, 5): 7
rand(Int64, 10): 10
rand(Int64, 20): 12
rand(Int64, 50): 20
rand(Int64, 100): 30
rand(Int64, 1000): 237
rand(Int64, 10000): 3230
rand(Int64, 100000): 43568
rand(Int64, 1000000): 757348
rand(Int64, 10000000): 8574296

julia> bench(Int, isqe)
ERROR: UndefVarError: isqe not defined
 in eval(::Module, ::Any) at ./boot.jl:225
 in macro expansion at ./REPL.jl:92 [inlined]
 in (::Base.REPL.##1#2{Base.REPL.REPLBackend})() at ./event.jl:46

OLD ==
rand(Int64, 0): 8
rand(Int64, 1): 9
rand(Int64, 2): 10
rand(Int64, 3): 11
rand(Int64, 4): 13
rand(Int64, 5): 15
rand(Int64, 10): 22
rand(Int64, 20): 38
rand(Int64, 50): 93
rand(Int64, 100): 171
rand(Int64, 1000): 1583
rand(Int64, 10000): 15700
rand(Int64, 100000): 156662
rand(Int64, 1000000): 1603819
rand(Int64, 10000000): 16082368
```

It is a bit concerning that the time taken for the new version seems slightly superlinear, but it is so much faster that the superlinearity doesn't even seem to matter.