@TotalVerb 2016-09-14
I've rebased this. This performance optimization should still be valid on v0.6.