@tkelman 2016-09-15
dunno if this is covered by existing benchmarks, but @nanosoldier `runbenchmarks(ALL, vs = ":master")`