@nanosoldier 2016-09-15
[Your benchmark job](https://github.com/JuliaLang/julia/pull/16877#issuecomment-247205277) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/6e67c95a0f2f5c603363449f07bf26e4173c69bd/93ca4c7_vs_fdf3748/report.md). cc @jrevels