@TotalVerb 2016-10-07
I can't confirm right now, but those regressions look like noise to me.